refactor: main app block for separate sections + add routing + about page
