+++ package.json
@@ -17,9 +17,11 @@
   },
   "dependencies": {
     "@tailwindcss/vite": "^4.1.11",
+    "clsx": "^2.1.1",
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
     "react-icons": "^5.5.0",
+    "react-router": "^7.7.0",
     "tailwindcss": "^4.1.11"
   },
   "devDependencies": {
