+++ src/App.tsx
@@ -1,87 +1,18 @@
-import { useEffect, useState } from "react";
-import { Search } from "./components/Search";
-import { ListOfCharacters } from "./components/ListOfCharacters";
-import { isResponse } from "./utils/checkFn/isResponse";
-import type { StateAppComponent } from "./models/types/stateAppComponent";
-import { requestAPI } from "./utils/requestAPI";
 import { ErrorBoundary } from "./components/ErrorBoundary";
-import { CrashComponent } from "./components/CrashComponent";
-import { loadDataFromLocalStorage } from "./utils/loadDataFromLocalStorage";
-import { LocalStorageKey } from "./models/enums/localStorageKey";
 import { Footer } from "./components/Footer";
+import { Outlet } from "react-router";
+import { NavigationBar } from "./components/NavigationBar";
 
 const App = () => {
-  const [state, setState] = useState<StateAppComponent>({
-    characters: [],
-    isLoading: false,
-    inputSearch: "",
-    isError: false,
-    responseStatus: null,
-  });
-
-  const handleSearchInputChange = (value: string): void => {
-    setState({ ...state, inputSearch: value });
-  };
-
-  const requestToApi = async (search: string | null = null): Promise<void> => {
-    try {
-      setState({ ...state, responseStatus: null, isLoading: true });
-      const response = await requestAPI(search || state.inputSearch);
-      if (response.status >= 400) {
-        setState({ ...state, responseStatus: response.status });
-      }
-      if (!response.ok) return;
-
-      const data = await response.json();
-      if (isResponse(data)) {
-        setState({ ...state, characters: data.characters, isLoading: false });
-      } else {
-        setState({ ...state, characters: [], isLoading: false });
-      }
-    } catch (error) {
-      console.error("Ошибка загрузки:", error);
-      setState({ ...state, characters: [], isLoading: false });
-    }
-  };
-
-  useEffect(() => {
-    const savedSearch = loadDataFromLocalStorage(LocalStorageKey.inputData);
-    const searchQuery = savedSearch ? String(savedSearch) : "";
-    setState((prev) => ({ ...prev, inputSearch: searchQuery }));
-    requestToApi(searchQuery);
-  }, [state.inputSearch]);
-
-  const { characters, isLoading, isError, inputSearch, responseStatus } = state;
-
   return (
     <div className="container mx-auto p-3 flex flex-col gap-3 h-full">
-      <div className="bg-blue-100 rounded-2xl p-3 flex flex-col gap-2 grow">
-        <h1 className="p-4 text-4xl">StarTrek characters library:</h1>
-        <ErrorBoundary>
-          <Search
-            onInputChange={handleSearchInputChange}
-            isLoading={isLoading}
-            initialValue={inputSearch}
-          />
-          {isError && <CrashComponent />}
-          {isLoading ? (
-            <p className="text-gray-500 text-xl" aria-label="Loading data...">
-              Loading data...
-            </p>
-          ) : (
-            <ListOfCharacters characters={characters} />
-          )}
-          {responseStatus !== null && (
-            <p
-              className="text-red-600 text-sm font-medium mt-2"
-              data-testid="error-message"
-            >
-              Error connecting to API. Status code: {responseStatus}
-            </p>
-          )}
-        </ErrorBoundary>
-      </div>
-      <Footer />
+      <ErrorBoundary>
+        <div className="bg-blue-100 rounded-2xl p-3 flex flex-col gap-2 grow">
+          <NavigationBar />
+          <Outlet />
+        </div>
+        <Footer />
+      </ErrorBoundary>
     </div>
   );
 };
