+++ src/constant/global-constant.ts
@@ -1,3 +1,4 @@
 export const LINK_TO_API = "https://stapi.co/api/v1/rest/character/search";
 export const LINK_TO_GITHUB = "https://github.com/kelzerock";
 export const LINK_TO_RS = "https://rs.school/";
+export const ELEMENTS_PER_PAGE = 20;
