+++ src/main.tsx
@@ -1,7 +1,8 @@
 import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 import "./index.css";
-import App from "./App.tsx";
+import { RouterProvider } from "react-router";
+import { router } from "./router/router.tsx";
 
 const root = document.createElement("div");
 root.id = "root";
@@ -10,6 +11,6 @@ document.body.append(root);
 
 createRoot(root).render(
   <StrictMode>
-    <App />
+    <RouterProvider router={router} />
   </StrictMode>,
 );
