+++ src/models/enums/paths.ts
@@ -1,0 +1,4 @@
+export const Paths = {
+  home: "/",
+  about: "about",
+} as const;
