+++ src/page/aboutPage/AboutPage.tsx
@@ -1,0 +1,20 @@
+import React from "react";
+
+const AboutPage = () => {
+  return (
+    <div>
+      <h1>About author:</h1>
+      <h2>Hi, I&apos;m Aleksei</h2>
+      <p>
+        Lorem ipsum dolor sit amet consectetur adipisicing elit. Et iure omnis
+        eligendi molestias nesciunt temporibus corrupti quisquam odio totam.
+        Accusantium.
+      </p>
+      <a href="https://rs.school/" target="_blank" rel="noreferrer">
+        The best school for web developer
+      </a>
+    </div>
+  );
+};
+
+export default AboutPage;
