+++ src/utils/requestAPI.ts
@@ -1,7 +1,7 @@
-import { LINK_TO_API } from "../constant/global-constant";
+import { ELEMENTS_PER_PAGE, LINK_TO_API } from "../constant/global-constant";
 
 export const requestAPI = async (name: string) => {
-  return await fetch(LINK_TO_API, {
+  return await fetch(`${LINK_TO_API}?pageSize=${ELEMENTS_PER_PAGE}`, {
     method: "POST",
     headers: {
       "Content-Type": "application/x-www-form-urlencoded; charset=UTF-8",
